Fixed #24: Long adressing everywhere

+++ core/src/main/java/com/yandex/yoctodb/util/mutable/BitSet.java
@@ -108,7 +108,7 @@ public interface BitSet {
     boolean or(
             @NotNull
             Buffer longArrayBitSetInByteBuffer,
-            int startPosition,
+            long startPosition,
             int bitSetSizeInLongs);
 
     /**
